+++ jdbc/src/main/java/nextstep/jdbc/JdbcTemplate.java
@@ -1,5 +1,6 @@
 package nextstep.jdbc;
 
+import nextstep.jdbc.exception.JdbcTemplateException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
