+++ jdbc/src/main/java/nextstep/jdbc/exception/JdbcTemplateException.java
@@ -1,4 +1,4 @@
-package nextstep.jdbc;
+package nextstep.jdbc.exception;
 
 import java.sql.SQLException;
