+++ learning/src/test/java/reflection/Junit3Test.java
@@ -1,15 +1,15 @@
 package reflection;
 
 public class Junit3Test {
-    public void test1() throws Exception {
+    public void test1() {
         System.out.println("Running Test1");
     }
 
-    public void test2() throws Exception {
+    public void test2() {
         System.out.println("Running Test2");
     }
 
-    public void three() throws Exception {
+    public void three() {
         System.out.println("Running Test3");
     }
 }
